print return from file and cell
